Divorce seed function from seed data
